Rename venue to location in schedule
A few folks have mentioned to me that the use of venue on the schedule
was confusing to them.  Generally venue refers to an overall building,
not a room within a building.  Some of the room names contributed to
this confusion, such as Stadium.  Location is a bit of a broader term
that is more applicable to room names.